requirements.txt
mysql-connector-python
pyodbc
Django
